requirements.txt
yt-dlp>=2024.1.1
google-genai>=1.0.0
python-dotenv>=1.0.0
